Create transient `Artifacts.toml` to manage `gc()` better.
Now that the `julia` buildkite plugin is calling `gc()` quite
aggressively, we need our rootfs images to look like actual, anchored
artifacts, otherwise they are cleaned up too aggressively.

By creating a (transient) `Artifacts.toml` file, we can better control
the generational nature of the Pkg GC.  If we set our collect delay to,
for instance, 2 days, and each time we use a rootfs we call
`bind_artifact!()` with the rootfs treehash, then we get the desirable
behavior that a rootfs is only deleted if it has not been used in the
last 2 days.